python_deployment/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
python-multipart==0.0.20
jinja2==3.1.6
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
alembic==1.16.4
python-dotenv==1.1.1